chore: move helpers into the state module
