+++ src/app/state/helper/binary-document.test.ts
