+++ src/app/state/helper/binary-document.ts
